Moved FinBud into frontend folder and removed Clerk
