+++ FinBud-Frontend/.eslintrc.json
