+++ FinBud-Frontend/next.config.mjs
