+++ FinBud-Frontend/package.json
@@ -11,6 +11,7 @@
     "deploy": "gh-pages -d out"
   },
   "dependencies": {
+    "@auth0/auth0-react": "^2.2.4",
     "@clerk/nextjs": "^5.6.3",
     "@tabler/icons-react": "^3.19.0",
     "bootstrap": "^5.3.3",
