+++ FinBud-Frontend/public/images/FinHappy.png
