+++ FinBud-Frontend/public/images/bg-lg.png
