+++ FinBud-Frontend/src/app/characters/page.tsx
