+++ FinBud-Frontend/src/app/favicon.ico
