+++ FinBud-Frontend/src/app/globals.css
