+++ FinBud-Frontend/src/app/loading.tsx
