+++ FinBud-Frontend/src/app/page.tsx
