+++ FinBud-Frontend/src/app/tips/page.tsx
