+++ FinBud-Frontend/src/components/Calculator.tsx
