+++ FinBud-Frontend/src/components/ResponsiveImage.tsx
