+++ FinBud-Frontend/src/components/Textbox.tsx
