+++ FinBud-Frontend/src/components/answer.tsx
