+++ FinBud-Frontend/src/data/easier.json
