+++ FinBud-Frontend/src/quiz-logic/quiz-functions.ts
