+++ FinBud-Frontend/src/types.d.ts
